v0.3.0

+++ src/UI/Component/WindowRECT.h
@@ -3,7 +3,7 @@
 #include <imgui.h>
 class WindowRECT {
 public:
-	static void i_ClampedLerp(RECT NewRECT, float l_Step, float Threshold);
+	static void i_ClampedLerp(RECT NewRECT, LONG l_Step, float Threshold);
 
 	static void i_Bisectio(RECT NewRECT, float Threshold);
